% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_filter_steps.R
\name{all_filter_steps}
\alias{all_filter_steps}
\title{This function performs all of the steps involved in filtering false-positive
detections from a Lotek JSATS receiver. The first few simpler filter steps
are coded directly in this function (e.g. multipath, minimum number of detections).
Then this function calls pri_filter() which performs the complex filter based
on the nominalPRI of the tag.}
\usage{
all_filter_steps(Lotek_input_file, settings = settings)
}
\description{
This function performs all of the steps involved in filtering false-positive
detections from a Lotek JSATS receiver. The first few simpler filter steps
are coded directly in this function (e.g. multipath, minimum number of detections).
Then this function calls pri_filter() which performs the complex filter based
on the nominalPRI of the tag.
}
