% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_filter_Lotek.R
\name{parallel_filter_Lotek}
\alias{parallel_filter_Lotek}
\title{filter multiple csv files that were pre-processed by parallel_raw_Lotek() or
process_single_raw() in parallel. Or manually reformat data files to match
required format (columns: DateTime, HexID, nominalPRI)
This is the main executable function that the user will deploy to filter data.
This function calls multiple helper functions that contain the actual filter
criteria. process_single_file() performs all of the processes necessary on
on a csv file that has been properly preformatted. parallel_filter_Lotek()
simply parallelizes the work that process_single_file() does.}
\usage{
parallel_filter_Lotek(
  input_files,
  input_prefix = NA,
  output_prefix = "Filtered_",
  output_path,
  allow_overwrite = FALSE,
  keep_rejected = FALSE,
  n_cores = max(1, round(parallel::detectCores()/2)),
  nominalPRI = NULL,
  pri_table = NULL,
  pri_tag_col = NULL,
  pri_value_col = NULL,
  detection_window = 16.6,
  min_detections = 4,
  sd_threshold = 0.025,
  multipath_threshold = 0.3,
  nominalPRI_threshold = 0.2
)
}
\description{
filter multiple csv files that were pre-processed by parallel_raw_Lotek() or
process_single_raw() in parallel. Or manually reformat data files to match
required format (columns: DateTime, HexID, nominalPRI)
This is the main executable function that the user will deploy to filter data.
This function calls multiple helper functions that contain the actual filter
criteria. process_single_file() performs all of the processes necessary on
on a csv file that has been properly preformatted. parallel_filter_Lotek()
simply parallelizes the work that process_single_file() does.
}
