% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pri_filter.R
\name{pri_filter}
\alias{pri_filter}
\title{This function filters the detection records based on the nominalPRI of the tag.
It searches for min_dets detections within a time window of nominalPRI * detection_window seconds.
The function considers all combinations of min_dets that may occur within the
time window to see if any of them pass the filter criteria. This is done
to avoid a false-positive inside a set of true-positive detections from
causing a violation (such as PRI standard deviation)}
\usage{
pri_filter(dets_df, settings = settings)
}
\description{
This function filters the detection records based on the nominalPRI of the tag.
It searches for min_dets detections within a time window of nominalPRI * detection_window seconds.
The function considers all combinations of min_dets that may occur within the
time window to see if any of them pass the filter criteria. This is done
to avoid a false-positive inside a set of true-positive detections from
causing a violation (such as PRI standard deviation)
}
