% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_raw_Lotek.R
\name{process_raw_Lotek_dets}
\alias{process_raw_Lotek_dets}
\title{parallel_raw_Lotek()
Function to process a directory of raw Lotek TXT files exported by WHS Host
batch file conversion tool. This function uses process_single_raw() as the primary
worker function but parallelizes it to use multiple cores to process multiple
input TXt files simulatneously.
NOTE: If input_path and output_path are the same and an output_prefix is not provided,
then the input files will be overwritten.}
\usage{
process_raw_Lotek_dets(
  input_path,
  raw_files = list.files(input_path, pattern = "\\\\.TXT$"),
  output_path,
  output_prefix = "",
  AllowableTagCodes = NULL,
  tz = "Etc/GMT+8",
  n_cores = max(1, round(parallel::detectCores()/2))
)
}
\description{
parallel_raw_Lotek()
Function to process a directory of raw Lotek TXT files exported by WHS Host
batch file conversion tool. This function uses process_single_raw() as the primary
worker function but parallelizes it to use multiple cores to process multiple
input TXt files simulatneously.
NOTE: If input_path and output_path are the same and an output_prefix is not provided,
then the input files will be overwritten.
}
