% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_single_file.R
\name{process_single_file}
\alias{process_single_file}
\title{This function reads, formats, and calls necessary filter functions to fully
process a single csv file that was exported by parallel_raw_Lotek() or
process_single_raw(). This is the function that parallel_filter_Lotek()
parallelizes to increase processing speed. This function calls multiple
helper functions which actually execute the filter criteria.}
\usage{
process_single_file(filepath, settings = settings)
}
\description{
This function reads, formats, and calls necessary filter functions to fully
process a single csv file that was exported by parallel_raw_Lotek() or
process_single_raw(). This is the function that parallel_filter_Lotek()
parallelizes to increase processing speed. This function calls multiple
helper functions which actually execute the filter criteria.
}
