% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_single_raw.R
\name{process_single_raw}
\alias{process_single_raw}
\title{process a single raw Lotek .TXT file exported by Lotek's WHS Host
batch file conversion tool.
This function takes .TXT files created by WHS Host's batch file conversion tool
which conversion .JST files from the receiver to unecrypted, usable text files.
This function reformats the TXT files output why WHS host and optionally
filters to only keep detections from a list of allowabe tag codes.
This is recommended as it will greatly reduce file size and thus processing.
This function and package are designed to be run each input file independently
in parallel due to the potential large file sizes which can cause memory limitations
if multiple files are retained in a single R environment.}
\usage{
process_single_raw(
  raw_file,
  output_path,
  output_prefix = "",
  AllowableTagCodes = NULL,
  tz = "Etc/GMT+8"
)
}
\description{
process a single raw Lotek .TXT file exported by Lotek's WHS Host
batch file conversion tool.
This function takes .TXT files created by WHS Host's batch file conversion tool
which conversion .JST files from the receiver to unecrypted, usable text files.
This function reformats the TXT files output why WHS host and optionally
filters to only keep detections from a list of allowabe tag codes.
This is recommended as it will greatly reduce file size and thus processing.
This function and package are designed to be run each input file independently
in parallel due to the potential large file sizes which can cause memory limitations
if multiple files are retained in a single R environment.
}
