% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_filter_steps.R
\name{all_filter_steps}
\alias{all_filter_steps}
\title{Broad filter function}
\usage{
all_filter_steps(Lotek_input_file, settings = settings)
}
\arguments{
\item{Lotek_input_file}{a reformatted Lotek detections file previously processed by parallel_raw_Lotek() or process_single_raw()}

\item{settings}{a list of filter settings passed from parallel_filter_Lotek()}
}
\value{
a filtered dataframe with all filter criteria applied
}
\description{
This function performs all of the steps involved in filtering false-positive
detections from a Lotek JSATS receiver. The first few simpler filter steps
are coded directly in this function (e.g. multipath, minimum number of detections).
Then this function calls pri_filter() which performs the complex filter based
on the nominalPRI of the tag.
}
